Split addEmployeeForm to other component
